Combined Main Tests
